hk-job-aggregator/requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.17.2
scrapy==2.11.0
lxml==5.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.2.0
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6

# Job Deduplication
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# API Enrichment (optional)
openai==1.10.0  # For AI-powered job descriptions/matching
